Refactored resolutions to be independent.
