+++ Native/Dependencies/Containers/DependencyContainer.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Chopsticks.Dependencies.Resolutions;
+using System;
 using System.Collections.Generic;
 
 namespace Chopsticks.Dependencies.Containers
@@ -6,81 +7,6 @@ namespace Chopsticks.Dependencies.Containers
     /// <inheritdoc cref="IDependencyContainer"/>
     public class DependencyContainer : IDependencyContainer, IDisposable
     {
-        /// <summary>
-        /// Resolves, and maintains resolved implementation instances of, the dependency 
-        /// specified by the provided dependency specification.
-        /// </summary>
-        /// <param name="contract">The contract that this resolution fulfills.</param>
-        /// <param name="factory">The factory that provides implementations for resolution.</param>
-        private abstract class Resolution(Type contract, 
-            Func<IDependencyContainer, object> factory) : 
-            IDisposable
-        {
-            public Func<IDependencyContainer, object> Factory { get; private init; } = factory;
-
-            // TODO :: Use the registration in place of the specification (reduce memory footprint).
-
-            public DependencyRegistration Registration { get; private init; } = new()
-            {
-                Contract = contract,
-            };
-
-
-            public abstract void Dispose();
-
-            public abstract object Get(IDependencyContainer container);
-        }
-
-        private class ContainedResolution(Type contract, 
-            Func<IDependencyContainer, object> factory) :
-            Resolution(contract, factory)
-        {
-            private object? _instance;
-
-
-            public override void Dispose()
-            {
-                if (_instance is IDisposable disposable)
-                    disposable.Dispose();
-
-                _instance = null;
-            }
-
-            public override object Get(IDependencyContainer container) =>
-                _instance ??= Factory(container);
-        }
-
-        private class SingletonResolution(Type contract, 
-            Func<IDependencyContainer, object> factory) :
-            Resolution(contract, factory)
-        {
-            private object? _instance;
-
-            public override void Dispose()
-            {
-                if (_instance is IDisposable disposable)
-                    disposable.Dispose();
-
-                _instance = null;
-
-            }
-
-            public override object Get(IDependencyContainer container) => 
-                _instance ??= Factory(container);
-        }
-
-        private class TransientResolution(Type contract, 
-            Func<IDependencyContainer, object> factory) :
-            Resolution(contract, factory)
-        {
-            public override void Dispose()
-            {
-            }
-
-            public override object Get(IDependencyContainer container) => Factory(container);
-        }
-
-
         /// <inheritdoc/>
         public bool InheritParentDependencies { get; set; } = true;
 
@@ -91,7 +17,10 @@ namespace Chopsticks.Dependencies.Containers
         /// <summary>
         /// A mapping of contract types to all resolutions for each contract.
         /// </summary>
-        private readonly Dictionary<Type, List<Resolution>> _resolutions = [];
+        private readonly Dictionary<Type, List<DependencyResolution>> _resolutions = [];
+
+
+        // TODO :: Leverage a resolution factory.
 
 
         /// <inheritdoc/>
@@ -120,7 +49,7 @@ namespace Chopsticks.Dependencies.Containers
             // TODO :: May need to handle Contained (inherited resolutions) differently.
 
             int index = 0;
-            Resolution? resolution = null;
+            DependencyResolution? resolution = null;
             var resolutions = _resolutions[registration.Contract];
             int count = resolutions.Count;
             for (; index < count; index++)
@@ -148,7 +77,8 @@ namespace Chopsticks.Dependencies.Containers
         public IDependencyContainer Register(DependencySpecification specification, 
             out DependencyRegistration registration)
         {
-            Resolution resolution = specification.Lifetime switch
+            // TODO :: Replace with injected factory.
+            DependencyResolution resolution = specification.Lifetime switch
             {
                 DependencyLifetime.Contained => new ContainedResolution(
                     specification.Contract, specification.ImplementationFactory),
@@ -181,11 +111,11 @@ namespace Chopsticks.Dependencies.Containers
         {
             implementation = null;
 
-            var resolution = GetResolution(dependencyType);
+            var resolution = this.GetResolution(dependencyType);
             if (resolution == null)
-                return false;
+                return Parent?.Resolve(dependencyType, out implementation) is true;
 
-
+            implementation = resolution.Get(this);
             return true;
         }
 
@@ -209,8 +139,11 @@ namespace Chopsticks.Dependencies.Containers
         /// implementation type or the contract type.</param>
         /// <returns>The first resolution for the specified dependency type, 
         /// or null if there is no such known resolution.</returns>
-        private Resolution? GetResolution(Type dependencyType)
+        private DependencyResolution? GetResolution(Type dependencyType)
         {
+            if (!_resolutions.TryGetValue(dependencyType, out var resolution))
+                return null;
+
             // TODO :: Find the local resolution first, then check parent's resolutions.
             // TODO :: If the resolution is Contained, cache it as a local resolution during resolving.
             return null;
@@ -222,7 +155,7 @@ namespace Chopsticks.Dependencies.Containers
         /// <param name="dependencyType">The type of the dependency, either as the 
         /// implementation type or the contract type.</param>
         /// <returns>All resolutions for the specified dependency type.</returns>
-        private IEnumerable<Resolution> GetResolutions(Type dependencyType)
+        private IEnumerable<DependencyResolution> GetResolutions(Type dependencyType)
         {
             // TODO :: Find the local resolution first, then join with unique parent's resolutions.
             // TODO :: If any resolution is Contained, cache it as a local resolution during resolving.
